Update di.rst
I did have a problem with Service Name Conventions. Name Conventions are things user has to follow for the in order for the code to work properly. Instead, we have here a list of built in services that occasionally point to a place that is not necessarily in sync with the name. I'm not sure how to restate it, maybe someone else could have a look at this section.