Add 'findFirstBy' to Models Doc
This method was missing from the docs. Plenty of examples spread
throughout the docs in a number of examples, but no clear documentation
demonstrating how to use it. Not until digging in the example apps did I
find it. Not sure if I explained it well enough. Feedback welcome, of
course.