Update 'string' filter type description
It took long time to get why quotes in request not escaped but encoded to html entity. So I have to check it in filter.zep:
> case Filter::FILTER_STRING:
> return filter_var(value, FILTER_SANITIZE_STRING);

But in description from http://php.net/manual/en/filter.filters.sanitize.php written:
"Strip tags, optionally strip or encode special characters." - so encoded and not escaped.