use Phalcon\Acl; was removed in v4
I removed "use Phalcon\Acl;" occurences because it's been deleted in v4. I replaced "use Phalcon\Acl;" by "use Phalcon\Acl\Enum;" in one place, to use it with the function setNoArgumentsDefaultAction(). In the description text, there are many references to Phalcon\Acl, I don't know if it's important to change it too.