Fixed/corrected documentation on DI & translations
Using only 'new Locale()' doesn't work, since it's only injectable object only, and it doesn't have an `_(..)` methods, it's also done the same with a template sample in the documentation. Maybe it's meant to be used somehow otherwise, but at least as it is now in the sample it's not working.