fix 'comparisson of Model with boolean' & 'error message cicle'
When comparing a Model Object, as in a $user or $product, for example, the return value of the query would be null or the object itself. So when comparing it with a boolean it would never enter the cicle. Changed it to compare with null.
Also fixed the error messages cicle, where it would iterate over Message Objects and give them to the flash->error, when it actually asks for a string. Added getMessage() to the Message object.