Error in example
The example is registering two different services: the declaration of the second one will not replace the first one: if this is the purpose, the proposed change fixes the issue. I am not sure about the change because the declaration of the example is not clear:
```When you request a service from the container, if it cannot be found by using the same name, it will try to load a class with the same name. This behavior allows you to replace any service with another, by simply registering a service with the common name```
So, to fall back to the autoloader, `AnotherComponet` doesn't match either `IndexController`...